requirements.txt
Flask==2.2.2
numpy
matplotlib
opencv-python
gunicorn
azure-storage-blob
python-dotenv